Create Dockerfile
Basing the file heavily on https://github.com/krmaxwell/maltrieve/blob/master/docker/Dockerfile. Making REMnux-specific stylistic tweaks. Avoiding the ENTRYPOINT tag to let the user specify custom command-line parameters. Explicitly installing "requests[security]" using PIP to get rid of the InsecureRequestWarning as outlined in http://stackoverflow.com/questions/29134512/insecureplatformwarning-a-true-sslcontext-object-is-not-available-this-prevent. Making it easier for the user to map the "archive" directory to get malware samples saved to the underlying host.