Update Ciphey Dockerfile
This change should have no impact to the usability of the container or command, but with a significant decrease in size (down from 1.95GB to 164MB).
- Use the Official Python 3.8 slim docker image
- Remove nonroot user from sudo group, and remove unneeded password
- Add ENTRYPOINT to simplify running the ciphey command
- Add additional example to comments for file and text inputs

If the version of Ciphey that is currently in remnux/ciphey is desired, `ciphey==5.2.0` can be specified in the pip3 command, but 5.2.0 does not decode the example text input.

Ciphey does not yet include a way to save potential decrypted output to file(s), so references to output files were removed. Standard redirects and pipes will return to the host, not the container. It also includes comments, not just the potentially decrypted final result.